add frag support to ENABLE_TX_PAGE_REUSE
Re-enabling frags means we can get better TCP performance again when
packets can be batched.

There are even more stability concerns with this commit, since frags
could be pretty large the DMA block can easily run out of RAM and crash
the system. There are some TODOs in the code noting what can be done so
this can be addressed in a follow up commit.

The downside is that jumbo frames will probably not be supported in
this commit yet. Also possible more compatibility issues since we now
need a 4MiB DMA block instead of 2MiB. Lastly the tx ring entries had
to be trimmed from 1024->256, however I haven't been able to find a
performance hit with this.