switch to an array of small dma tx allocations
Since the largest continuous block of RAM to allocate from the kernel
is 4MiB switching to pre-allocating this memory per ringentry. Since
the skb size can be large, seen it up to 30KiB, andcurrently don't know
its max size, allocating 64KiB per ring entry.This does mean we are
allocating 64MiB up front but this should bemostly safe until we can
find out how to calculate this.

The performance seems to be about 3%-5% less on 4 cores sending this way
1.62Mpps vs 1.69Mpps.

Also there still seem to be some random lockups when the system is idle.